refactored rest-interfaces
