+++ openvalidation-rest/src/main/java/io/openvalidation/rest/model/dto/astDTO/MainNode.java
@@ -24,12 +24,10 @@ import java.util.List;
 public class MainNode {
   private List<Variable> declarations;
   private List<GenericNode> scopes;
-  private List<OpenValidationExceptionDTO> errors;
 
   public MainNode() {
     this.declarations = new ArrayList<>();
     this.scopes = new ArrayList<>();
-    this.errors = new ArrayList<>();
   }
 
   public List<Variable> getDeclarations() {
@@ -50,13 +48,5 @@ public class MainNode {
 
   public void addScopes(List<GenericNode> scopes) {
     this.scopes.addAll(scopes);
-  }
-
-  public List<OpenValidationExceptionDTO> getErrors() {
-    return errors;
-  }
-
-  public void addError(OpenValidationExceptionDTO error) {
-    this.errors.add(error);
   }
 }
